+++ webrtc/src/webrtc_c/api/video/nv12_buffer.cc
@@ -37,6 +37,16 @@ WEBRTC_EXPORT int webrtc_NV12Buffer_height(
   auto buf = reinterpret_cast<const webrtc::NV12Buffer*>(self);
   return buf->height();
 }
+WEBRTC_EXPORT int webrtc_NV12Buffer_chroma_width(
+    const struct webrtc_NV12Buffer* self) {
+  auto buf = reinterpret_cast<const webrtc::NV12Buffer*>(self);
+  return (buf->width() + 1) / 2;
+}
+WEBRTC_EXPORT int webrtc_NV12Buffer_chroma_height(
+    const struct webrtc_NV12Buffer* self) {
+  auto buf = reinterpret_cast<const webrtc::NV12Buffer*>(self);
+  return (buf->height() + 1) / 2;
+}
 WEBRTC_EXPORT uint8_t* webrtc_NV12Buffer_MutableDataY(
     struct webrtc_NV12Buffer* self) {
   auto buf = reinterpret_cast<webrtc::NV12Buffer*>(self);